# ui/openapi/fitnessclub.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FitnessClub.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_List",
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_Schedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/mine": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_Mine",
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/cancel": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_Cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/sessions/{id}/bookings": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_Book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/bookings/{clientId}/cancel": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "operationId": "Sessions_CancelBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingResponse": {
        "required": [
          "clientId",
          "bookedAt",
          "cancelledAt"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "bookedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "BookSessionRequest": {
        "required": [
          "clientId"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "ScheduleSessionRequest": {
        "required": [
          "title",
          "type",
          "trainerId",
          "roomId",
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionType"
              }
            ]
          },
          "trainerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "roomId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "start": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "end": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "capacity": {
            "maximum": 500,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SessionResponse": {
        "required": [
          "id",
          "title",
          "type",
          "trainerId",
          "roomId",
          "start",
          "end",
          "capacity",
          "status",
          "activeBookingCount",
          "bookings"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "trainerId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "activeBookingCount": {
            "type": "integer",
            "format": "int32"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookingResponse"
            }
          }
        }
      },
      "SessionStatus": {
        "enum": [
          "Scheduled",
          "Cancelled"
        ]
      },
      "SessionSummaryResponse": {
        "required": [
          "id",
          "title",
          "type",
          "trainerId",
          "roomId",
          "start",
          "end",
          "capacity",
          "status",
          "activeBookingCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "trainerId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "activeBookingCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SessionType": {
        "enum": [
          "Group",
          "Individual"
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Auth0 access token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}